Owncloud 8.2.7

+++ public/apps/files_sharing/tests/sharedstorage.php
@@ -6,6 +6,7 @@
  * @author Robin Appelman <icewind@owncloud.com>
  * @author Thomas Müller <thomas.mueller@tmit.eu>
  * @author Vincent Petry <pvince81@owncloud.com>
+ * @author Markus Goetz <markus@woboq.com>
  *
  * @copyright Copyright (c) 2015, ownCloud, Inc.
  * @license AGPL-3.0
@@ -47,8 +48,10 @@ class Test_Files_Sharing_Storage extends OCA\Files_sharing\Tests\TestCase {
 	}
 
 	protected function tearDown() {
-		$this->view->unlink($this->folder);
-		$this->view->unlink($this->filename);
+		if ($this->view) {
+			$this->view->unlink($this->folder);
+			$this->view->unlink($this->filename);
+		}
 
 		\OC\Files\Filesystem::getLoader()->removeStorageWrapper('oc_trashbin');
 
@@ -85,8 +88,9 @@ class Test_Files_Sharing_Storage extends OCA\Files_sharing\Tests\TestCase {
 		$this->assertFalse($user2View->is_dir($this->folder));
 
 		// delete the local folder
-		$fullPath = \OC_Config::getValue('datadirectory') . '/' . self::TEST_FILES_SHARING_API_USER2 . '/files/localfolder';
-		rmdir($fullPath);
+		/** @var \OC\Files\Storage\Storage $storage */
+		list($storage, $internalPath)  = \OC\Files\Filesystem::resolvePath('/' . self::TEST_FILES_SHARING_API_USER2 . '/files/localfolder');
+		$storage->rmdir($internalPath);
 
 		//enforce reload of the mount points
 		self::loginHelper(self::TEST_FILES_SHARING_API_USER2);
@@ -440,5 +444,70 @@ class Test_Files_Sharing_Storage extends OCA\Files_sharing\Tests\TestCase {
 
 		self::loginHelper(self::TEST_FILES_SHARING_API_USER1);
 		$this->view->unlink($this->folder);
+	}
+
+	public function testNameConflict() {
+		self::loginHelper(self::TEST_FILES_SHARING_API_USER1);
+		$view1 = new \OC\Files\View('/' . self::TEST_FILES_SHARING_API_USER1 . '/files');
+		$view1->mkdir('foo');
+		$folderInfo1 = $view1->getFileInfo('foo');
+
+		self::loginHelper(self::TEST_FILES_SHARING_API_USER3);
+		$view3 = new \OC\Files\View('/' . self::TEST_FILES_SHARING_API_USER3 . '/files');
+		$view3->mkdir('foo');
+		$folderInfo2 = $view3->getFileInfo('foo');
+
+		// share a folder with the same name from two different users to the same user
+		self::loginHelper(self::TEST_FILES_SHARING_API_USER1);
+
+		\OCP\Share::shareItem('folder', $folderInfo1['fileid'], \OCP\Share::SHARE_TYPE_GROUP,
+			self::TEST_FILES_SHARING_API_GROUP1, 31);
+
+		self::loginHelper(self::TEST_FILES_SHARING_API_USER2);
+
+		self::loginHelper(self::TEST_FILES_SHARING_API_USER3);
+
+		\OCP\Share::shareItem('folder', $folderInfo2['fileid'], \OCP\Share::SHARE_TYPE_GROUP,
+			self::TEST_FILES_SHARING_API_GROUP1, 31);
+
+		self::loginHelper(self::TEST_FILES_SHARING_API_USER2);
+		$view2 = new \OC\Files\View('/' . self::TEST_FILES_SHARING_API_USER2 . '/files');
+
+		$this->assertTrue($view2->file_exists('/foo'));
+		$this->assertTrue($view2->file_exists('/foo (2)'));
+
+		$mount = $view2->getMount('/foo');
+		$this->assertInstanceOf('\OCA\Files_Sharing\SharedMount', $mount);
+		/** @var \OC\Files\Storage\Shared $storage */
+		$storage = $mount->getStorage();
+
+		$source = $storage->getFile('');
+		$this->assertEquals(self::TEST_FILES_SHARING_API_USER1, $source['uid_owner']);
+	}
+
+	public function testLongLock() {
+		// https://github.com/owncloud/core/issues/25376
+		$fn = str_repeat("x",250); // maximum name length in oc_filecache
+		self::loginHelper(self::TEST_FILES_SHARING_API_USER1);
+		$view1 = new \OC\Files\View('/' . self::TEST_FILES_SHARING_API_USER1 . '/files');
+		$view1->mkdir($fn);
+		$view1->mkdir($fn.'/'.'vincent');
+		$view1->file_put_contents($fn . '/vincent/' . $fn, "dummy file data\n");
+
+		$fileinfo = $view1->getFileInfo($fn);
+		$result = \OCP\Share::shareItem('folder', $fileinfo['fileid'], \OCP\Share::SHARE_TYPE_USER, self::TEST_FILES_SHARING_API_USER2, \OCP\Constants::PERMISSION_ALL);
+		$this->assertTrue($result);
+
+		self::loginHelper(self::TEST_FILES_SHARING_API_USER2);
+		$user2View = new \OC\Files\View('/' . self::TEST_FILES_SHARING_API_USER2 . '/files');
+		$this->assertTrue($user2View->file_exists($fn . '/vincent/' . $fn));
+
+		list($sharedStorage,$bla) = $user2View->resolvePath($fn . '/vincent/' . $fn);
+		$cache = $sharedStorage->getCache();
+		$scanner = $sharedStorage->getScanner();
+		$scanner->scan('');
+		$this->assertTrue($cache->inCache('/vincent'));
+		$this->assertTrue($cache->inCache('/vincent/'.$fn));
+		$this->assertEquals(16, $cache->get('/vincent/'.$fn)['size']);
 	}
 }
